Link entity models together
