+++ src/routes/assets/models.rs
@@ -28,16 +28,16 @@ pub struct Model {
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
 pub enum Relation {
     #[sea_orm(
-        belongs_to = "super::experiments::Entity",
+        belongs_to = "crate::routes::experiments::models::Entity",
         from = "Column::ExperimentId",
-        to = "super::experiments::Column::Id",
+        to = "crate::routes::experiments::models::Column::Id",
         on_update = "NoAction",
         on_delete = "NoAction"
     )]
     Experiments,
 }
 
-impl Related<super::experiments::Entity> for Entity {
+impl Related<crate::routes::experiments::models::Entity> for Entity {
     fn to() -> RelationDef {
         Relation::Experiments.def()
     }
