--- src/routes/campaigns/mod.rs
@@ -1,5 +1,0 @@
-pub mod models;
-pub mod views;
-
-#[cfg(test)]
-pub mod tests;
