--- src/routes/campaigns/tests.rs
@@ -1,984 +1,0 @@
-use crate::config::test_helpers::setup_test_app;
-use axum::body::{Body, to_bytes};
-use axum::http::{Request, StatusCode};
-use serde_json::{Value, json};
-use tower::ServiceExt;
-
-async fn extract_response_body(response: axum::response::Response) -> (StatusCode, Value) {
-    let status = response.status();
-    let bytes = to_bytes(response.into_body(), usize::MAX)
-        .await
-        .expect("Failed to read response body");
-    let body: Value = serde_json::from_slice(&bytes)
-        .unwrap_or_else(|_| json!({"error": "Invalid JSON response"}));
-    (status, body)
-}
-
-async fn create_test_project(app: &axum::Router) -> uuid::Uuid {
-    let project_data = json!({
-        "name": "Test Project",
-        "note": "Test project for location tests",
-        "colour": "#FF0000"
-    });
-
-    let response = app
-        .clone()
-        .oneshot(
-            Request::builder()
-                .method("POST")
-                .uri("/api/projects")
-                .header("content-type", "application/json")
-                .body(Body::from(project_data.to_string()))
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-
-    let (status, body) = extract_response_body(response).await;
-
-    assert!(body.is_object(), "Expected JSON object response");
-    assert_eq!(
-        status,
-        StatusCode::CREATED,
-        "Failed to create test project: {body:?}"
-    );
-
-    uuid::Uuid::parse_str(body["id"].as_str().unwrap()).unwrap()
-}
-
-#[tokio::test]
-async fn test_location_crud_operations() {
-    let app = setup_test_app().await;
-
-    // Create a test project for the location
-    let project_id = create_test_project(&app).await;
-
-    // Test creating a location with unique name
-    let location_data = json!({
-        "name": format!("Test Location API {}", uuid::Uuid::new_v4()),
-        "comment": "Location created via API test",
-        "project_id": project_id
-    });
-    let response = app
-        .clone()
-        .oneshot(
-            Request::builder()
-                .method("POST")
-                .uri("/api/locations")
-                .header("content-type", "application/json")
-                .body(Body::from(location_data.to_string()))
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-
-    let (status, body) = extract_response_body(response).await;
-    assert_eq!(
-        status,
-        StatusCode::CREATED,
-        "Failed to create location: {body:?}"
-    );
-
-    let location_id = body["id"].as_str().unwrap();
-
-    // Test reading the created location
-    let get_response = app
-        .clone()
-        .oneshot(
-            Request::builder()
-                .method("GET")
-                .uri(format!("/api/locations/{location_id}"))
-                .body(Body::empty())
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-
-    let (get_status, get_body) = extract_response_body(get_response).await;
-    assert_eq!(get_status, StatusCode::OK, "Failed to get location");
-    assert_eq!(get_body["id"], location_id);
-
-    // Test updating the location
-    let update_data = json!({
-        "name": format!("Updated Location {}", uuid::Uuid::new_v4()),
-        "comment": "Updated via API test",
-        "project_id": project_id
-    });
-
-    let update_response = app
-        .clone()
-        .oneshot(
-            Request::builder()
-                .method("PUT")
-                .uri(format!("/api/locations/{location_id}"))
-                .header("content-type", "application/json")
-                .body(Body::from(update_data.to_string()))
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-
-    let (update_status, update_body) = extract_response_body(update_response).await;
-    assert_eq!(
-        update_status,
-        StatusCode::OK,
-        "Failed to update location: {update_body:?}"
-    );
-
-    // Test deleting the location
-    let delete_response = app
-        .clone()
-        .oneshot(
-            Request::builder()
-                .method("DELETE")
-                .uri(format!("/api/locations/{location_id}"))
-                .body(Body::empty())
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-
-    let (delete_status, _) = extract_response_body(delete_response).await;
-    assert_eq!(delete_status, StatusCode::NO_CONTENT);
-
-    // Test listing locations
-    let list_response = app
-        .clone()
-        .oneshot(
-            Request::builder()
-                .method("GET")
-                .uri("/api/locations")
-                .body(Body::empty())
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-
-    let (list_status, list_body) = extract_response_body(list_response).await;
-    assert_eq!(list_status, StatusCode::OK, "Failed to get locations");
-    assert!(
-        list_body.is_array(),
-        "Locations list should be a direct array"
-    );
-}
-
-#[tokio::test]
-async fn test_location_validation() {
-    let app = setup_test_app().await;
-
-    // Test creating location with invalid data (null name)
-    let invalid_data = json!({
-        "name": null,
-        "comment": "Invalid location"
-    });
-
-    let response = app
-        .clone()
-        .oneshot(
-            Request::builder()
-                .method("POST")
-                .uri("/api/locations")
-                .header("content-type", "application/json")
-                .body(Body::from(invalid_data.to_string()))
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-
-    let (status, _body) = extract_response_body(response).await;
-    assert!(
-        status.is_client_error(),
-        "Should reject location with null name"
-    );
-
-    // Test creating location with missing required fields
-    let incomplete_data = json!({
-        "comment": "Incomplete location"
-        // Missing name
-    });
-
-    let response = app
-        .clone()
-        .oneshot(
-            Request::builder()
-                .method("POST")
-                .uri("/api/locations")
-                .header("content-type", "application/json")
-                .body(Body::from(incomplete_data.to_string()))
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-
-    let (status, _body) = extract_response_body(response).await;
-    assert!(
-        status.is_client_error(),
-        "Should reject incomplete location data"
-    );
-}
-
-#[tokio::test]
-async fn test_location_filtering_and_pagination() {
-    let app = setup_test_app().await;
-
-    // Create a test project for the locations
-    let project_id = create_test_project(&app).await;
-
-    // Create some test locations for filtering
-    for i in 1..=3 {
-        let location_data = json!({
-            "name": format!("Filter Test Location {}", i),
-            "comment": format!("Test location {} for filtering", i),
-            "project_id": project_id
-        });
-
-        let response = app
-            .clone()
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/api/locations")
-                    .header("content-type", "application/json")
-                    .body(Body::from(location_data.to_string()))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        let (status, _) = extract_response_body(response).await;
-        assert_eq!(status, StatusCode::CREATED);
-    }
-
-    // Test filtering by project_id
-    let filter_response = app
-        .clone()
-        .oneshot(
-            Request::builder()
-                .method("GET")
-                .uri(format!("/api/locations?filter[project_id]={project_id}"))
-                .body(Body::empty())
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-
-    let (filter_status, filter_body) = extract_response_body(filter_response).await;
-    assert_eq!(
-        filter_status,
-        StatusCode::OK,
-        "Failed to filter locations by project_id"
-    );
-    let items = filter_body.as_array().unwrap();
-    assert!(items.len() >= 3, "Should find at least 3 locations");
-
-    // Test React-Admin pagination
-    let page_response = app
-        .clone()
-        .oneshot(
-            Request::builder()
-                .method("GET")
-                .uri("/api/locations?range=[0,1]")
-                .body(Body::empty())
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-
-    let (page_status, page_body) = extract_response_body(page_response).await;
-    assert_eq!(page_status, StatusCode::OK, "Failed to paginate locations");
-
-    let paginated_items = page_body.as_array().unwrap().len();
-    assert!(
-        paginated_items == 2,
-        "Should return 2 items. returned: {paginated_items}"
-    );
-}
-
-// Helper function to create a test location
-async fn create_test_location(
-    app: &axum::Router,
-    project_id: &str,
-) -> Result<(String, serde_json::Value), String> {
-    let location_data = json!({
-        "name": format!("Test Location CRUD {}", uuid::Uuid::new_v4()),
-        "comment": "Test location for CRUD operations",
-        "project_id": project_id
-    });
-
-    let response = app
-        .clone()
-        .oneshot(
-            Request::builder()
-                .method("POST")
-                .uri("/api/locations")
-                .header("content-type", "application/json")
-                .body(Body::from(location_data.to_string()))
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-
-    let (status, body) = extract_response_body(response).await;
-
-    if status == StatusCode::CREATED {
-        // Validate response structure - this assertion serves as success confirmation
-        assert!(
-            body["id"].is_string(),
-            "Location creation successful - Response should include ID"
-        );
-        assert!(
-            body["name"]
-                .as_str()
-                .unwrap()
-                .contains("Test Location CRUD")
-        );
-        assert_eq!(body["comment"], "Test location for CRUD operations");
-        assert!(body["created_at"].is_string());
-        assert!(body["last_updated"].is_string());
-
-        let location_id = body["id"].as_str().unwrap().to_string();
-        Ok((location_id, body))
-    } else {
-        Err(format!(
-            "Location creation failed: Status {status}, Body: {body}"
-        ))
-    }
-}
-
-// Helper function to test location retrieval
-async fn test_location_retrieval(app: &axum::Router, location_id: &str) {
-    let get_response = app
-        .clone()
-        .oneshot(
-            Request::builder()
-                .method("GET")
-                .uri(format!("/api/locations/{location_id}"))
-                .body(Body::empty())
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-
-    let (get_status, get_body) = extract_response_body(get_response).await;
-    if get_status == StatusCode::OK {
-        // This assertion confirms successful retrieval
-        assert_eq!(
-            get_body["id"], location_id,
-            "Location retrieval successful - ID should match"
-        );
-        assert!(
-            get_body["name"]
-                .as_str()
-                .unwrap()
-                .contains("Test Location CRUD")
-        );
-
-        // Validate related data structure
-        if get_body["experiments"].is_array() {
-            println!("   ✅ Experiments array present");
-        }
-        if get_body["samples"].is_array() {
-            println!("   ✅ Samples array present");
-        }
-    } else {
-        println!("⚠️  Location retrieval failed: {get_status}");
-    }
-}
-
-// Helper function to test location update
-async fn test_location_update(app: &axum::Router, location_id: &str) -> bool {
-    let update_data = json!({
-        "comment": "Updated comment for location"
-    });
-
-    let update_response = app
-        .clone()
-        .oneshot(
-            Request::builder()
-                .method("PATCH")
-                .uri(format!("/api/locations/{location_id}"))
-                .header("content-type", "application/json")
-                .body(Body::from(update_data.to_string()))
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-
-    let (update_status, update_body) = extract_response_body(update_response).await;
-    match update_status {
-        StatusCode::OK => {
-            // This assertion confirms successful update
-            assert_eq!(
-                update_body["comment"], "Updated comment for location",
-                "Location update successful - Comment should be updated"
-            );
-            true
-        }
-        StatusCode::METHOD_NOT_ALLOWED => {
-            println!("⚠️  Location update not implemented (405) - This is expected");
-            false
-        }
-        _ => {
-            println!("📋 Location update returned: {update_status}");
-            false
-        }
-    }
-}
-
-// Helper function to test location deletion
-async fn test_location_deletion(app: &axum::Router, location_id: &str) -> bool {
-    let delete_response = app
-        .clone()
-        .oneshot(
-            Request::builder()
-                .method("DELETE")
-                .uri(format!("/api/locations/{location_id}"))
-                .body(Body::empty())
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-
-    let delete_status = delete_response.status();
-    match delete_status {
-        status if status.is_success() => {
-            println!("✅ Location delete successful");
-            true
-        }
-        StatusCode::METHOD_NOT_ALLOWED => {
-            println!("⚠️  Location delete not implemented (405) - This is expected");
-            false
-        }
-        StatusCode::NOT_FOUND => {
-            println!(
-                "📋 Location delete returned 404 (location not found) - This is expected for non-existent resources"
-            );
-            false
-        }
-        _ => {
-            println!("📋 Location delete returned: {delete_status}");
-            false
-        }
-    }
-}
-
-#[tokio::test]
-async fn test_location_list_operations() {
-    let app = setup_test_app().await;
-
-    // Test getting all locations
-    let list_response = app
-        .clone()
-        .oneshot(
-            Request::builder()
-                .method("GET")
-                .uri("/api/locations")
-                .body(Body::empty())
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-
-    let (list_status, list_body) = extract_response_body(list_response).await;
-
-    if list_status == StatusCode::OK {
-        // This assertion confirms successful listing
-        assert!(
-            list_body.is_array(),
-            "Location listing successful - Response should be an array"
-        );
-        let locations = list_body.as_array().unwrap();
-
-        // Validate structure of locations in list
-        for location in locations {
-            assert!(location["id"].is_string(), "Each location should have ID");
-            assert!(
-                location["name"].is_string(),
-                "Each location should have name"
-            );
-            assert!(
-                location["created_at"].is_string(),
-                "Each location should have created_at"
-            );
-            assert!(
-                location["last_updated"].is_string(),
-                "Each location should have last_updated"
-            );
-        }
-    } else {
-        println!("⚠️  Location listing failed: Status {list_status}");
-        assert!(
-            list_status.is_client_error() || list_status.is_server_error(),
-            "Location listing should either succeed or fail gracefully"
-        );
-    }
-}
-
-#[tokio::test]
-async fn test_location_filtering_and_sorting() {
-    let app = setup_test_app().await;
-
-    // Create test locations for filtering
-    let test_locations = [
-        ("Filter Test Location A", "Project Alpha"),
-        ("Filter Test Location B", "Project Beta"),
-        ("Filter Test Location C", "Project Alpha"),
-    ];
-
-    let mut created_ids = Vec::new();
-
-    for (name, comment) in test_locations {
-        let location_data = json!({
-            "name": format!("{} {}", name, &uuid::Uuid::new_v4().to_string()[..8]),
-            "comment": comment
-        });
-
-        let response = app
-            .clone()
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/api/locations")
-                    .header("content-type", "application/json")
-                    .body(Body::from(location_data.to_string()))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        let (status, body) = extract_response_body(response).await;
-        if status == StatusCode::CREATED {
-            created_ids.push(body["id"].as_str().unwrap().to_string());
-        }
-    }
-
-    if created_ids.is_empty() {
-        // No test locations created - skip filtering tests
-    } else {
-        // Test filtering by comment
-        let filter_response = app
-            .clone()
-            .oneshot(
-                Request::builder()
-                    .method("GET")
-                    .uri("/api/locations?filter[comment]=Project%20Alpha")
-                    .body(Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        let (filter_status, filter_body) = extract_response_body(filter_response).await;
-
-        if filter_status == StatusCode::OK {
-            let filtered_locations = filter_body.as_array().unwrap();
-
-            // Check if filtering actually works (document known issue)
-            let mut non_matching_count = 0;
-            for location in filtered_locations {
-                if location["comment"] != "Project Alpha" {
-                    non_matching_count += 1;
-                    // KNOWN ISSUE: Filtering returned non-matching location
-                }
-            }
-
-            if non_matching_count == 0 && !filtered_locations.is_empty() {
-                println!("✅ Location filtering appears to work correctly");
-            } else if filtered_locations.is_empty() {
-                println!(
-                    "📋 Location filtering returned no results (may be working or not implemented)"
-                );
-            } else {
-                println!(
-                    "📋 KNOWN ISSUE: Location filtering returned {} non-matching results out of {} total",
-                    non_matching_count,
-                    filtered_locations.len()
-                );
-            }
-        } else {
-            println!("⚠️  Location filtering failed: Status {filter_status}");
-        }
-
-        // Test sorting by name
-        let sort_response = app
-            .clone()
-            .oneshot(
-                Request::builder()
-                    .method("GET")
-                    .uri("/api/locations?sort[name]=asc")
-                    .body(Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        let (sort_status, _) = extract_response_body(sort_response).await;
-
-        if sort_status == StatusCode::OK {
-            println!("✅ Location sorting endpoint accessible");
-        } else {
-            println!("⚠️  Location sorting failed: Status {sort_status}");
-        }
-    }
-}
-
-#[tokio::test]
-async fn test_location_not_found() {
-    let app = setup_test_app().await;
-
-    // Test getting non-existent location
-    let fake_id = uuid::Uuid::new_v4();
-    let response = app
-        .clone()
-        .oneshot(
-            Request::builder()
-                .method("GET")
-                .uri(format!("/api/locations/{fake_id}"))
-                .body(Body::empty())
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-
-    let (status, _body) = extract_response_body(response).await;
-    assert_eq!(
-        status,
-        StatusCode::NOT_FOUND,
-        "Should return 404 for non-existent location"
-    );
-}
-
-#[tokio::test]
-async fn test_location_project_assignment() {
-    let app = setup_test_app().await;
-
-    // Create a project first
-    let project_id = create_test_project(&app).await;
-
-    // Create location with project assignment
-    let location_data = json!({
-        "name": format!("Location with Project {}", uuid::Uuid::new_v4()),
-        "comment": "Testing project assignment",
-        "project_id": project_id
-    });
-
-    let response = app
-        .clone()
-        .oneshot(
-            Request::builder()
-                .method("POST")
-                .uri("/api/locations")
-                .header("content-type", "application/json")
-                .body(Body::from(location_data.to_string()))
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-
-    let (status, body) = extract_response_body(response).await;
-
-    if status == StatusCode::CREATED {
-        println!("✅ Location with project assignment created successfully");
-        assert_eq!(
-            uuid::Uuid::parse_str(body["project_id"].as_str().unwrap()).unwrap(),
-            project_id
-        );
-
-        // Test that the location can be retrieved with project info
-        let location_id = body["id"].as_str().unwrap();
-        let get_response = app
-            .clone()
-            .oneshot(
-                Request::builder()
-                    .method("GET")
-                    .uri(format!("/api/locations/{location_id}"))
-                    .body(Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        let (get_status, get_body) = extract_response_body(get_response).await;
-        if get_status == StatusCode::OK {
-            println!("✅ Location with project retrieved successfully");
-            assert_eq!(
-                uuid::Uuid::parse_str(get_body["project_id"].as_str().unwrap()).unwrap(),
-                project_id
-            );
-        }
-    } else {
-        println!("📋 Location with project assignment failed: Status {status}");
-    }
-}
-
-#[tokio::test]
-async fn test_location_related_data_structure() {
-    let app = setup_test_app().await;
-
-    // Create a simple location to test related data structure
-    let location_data = json!({
-        "name": format!("Related Data Test {}", uuid::Uuid::new_v4()),
-        "comment": "Testing related data loading"
-    });
-
-    let response = app
-        .clone()
-        .oneshot(
-            Request::builder()
-                .method("POST")
-                .uri("/api/locations")
-                .header("content-type", "application/json")
-                .body(Body::from(location_data.to_string()))
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-
-    let (status, body) = extract_response_body(response).await;
-
-    if status == StatusCode::CREATED {
-        let location_id = body["id"].as_str().unwrap();
-
-        // Get the location and check the related data structure
-        let get_response = app
-            .clone()
-            .oneshot(
-                Request::builder()
-                    .method("GET")
-                    .uri(format!("/api/locations/{location_id}"))
-                    .body(Body::empty())
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        let (get_status, get_body) = extract_response_body(get_response).await;
-
-        if get_status == StatusCode::OK {
-            println!("✅ Location related data structure test");
-
-            // Check that experiments array is present
-            if get_body["experiments"].is_array() {
-                let experiments = get_body["experiments"].as_array().unwrap();
-                println!(
-                    "   ✅ Experiments array present ({} items)",
-                    experiments.len()
-                );
-            } else {
-                println!("   ⚠️  Experiments array missing or wrong type");
-            }
-
-            // Check that samples array is present
-            if get_body["samples"].is_array() {
-                let samples = get_body["samples"].as_array().unwrap();
-                println!("   ✅ Samples array present ({} items)", samples.len());
-
-                // Check sample structure if samples exist
-                for sample in samples {
-                    if sample["treatments"].is_array() {
-                        println!("   ✅ Sample treatments array present");
-                    }
-                }
-            } else {
-                println!("   ⚠️  Samples array missing or wrong type");
-            }
-
-            println!("   📋 Related data loading appears to be working");
-        } else {
-            println!("📋 Could not test related data - location retrieval failed: {get_status}");
-        }
-    } else {
-        println!("📋 Skipping related data test - couldn't create location");
-    }
-}
-
-#[tokio::test]
-async fn test_location_complex_queries() {
-    let app = setup_test_app().await;
-
-    println!("📋 LOCATION COMPLEX QUERIES TEST");
-    println!("   Testing complex location query scenarios");
-
-    // Test pagination
-    let pagination_response = app
-        .clone()
-        .oneshot(
-            Request::builder()
-                .method("GET")
-                .uri("/api/locations?limit=5&offset=0")
-                .body(Body::empty())
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-
-    let (pagination_status, pagination_body) = extract_response_body(pagination_response).await;
-
-    if pagination_status == StatusCode::OK {
-        println!("   ✅ Pagination query successful");
-        let locations = pagination_body.as_array().unwrap();
-        println!("   Returned {} locations with limit=5", locations.len());
-    } else {
-        println!("   ⚠️  Pagination query failed: {pagination_status}");
-    }
-
-    // Test multiple filters
-    let multi_filter_response = app
-        .clone()
-        .oneshot(
-            Request::builder()
-                .method("GET")
-                .uri("/api/locations?filter[name]=Test&sort[created_at]=desc")
-                .body(Body::empty())
-                .unwrap(),
-        )
-        .await
-        .unwrap();
-
-    let (multi_filter_status, _) = extract_response_body(multi_filter_response).await;
-
-    if multi_filter_status == StatusCode::OK {
-        println!("   ✅ Multi-filter query successful");
-    } else {
-        println!("   ⚠️  Multi-filter query failed: {multi_filter_status}");
-    }
-
-    // This test always passes - it's for documenting query capabilities
-    // Documents location query behavior
-}
-
-// ===== TESTS USING PREVIOUSLY UNUSED HELPER FUNCTIONS =====
-
-#[tokio::test]
-async fn test_location_complete_lifecycle() {
-    let app = setup_test_app().await;
-
-    // Create a test project first
-    let project_id = create_test_project(&app).await;
-    let project_id_str = project_id.to_string();
-
-    // Use the unused helper function to create a location
-    let location_result = create_test_location(&app, &project_id_str).await;
-
-    match location_result {
-        Ok((location_id, _body)) => {
-            println!("✅ Location created using helper function: {location_id}");
-
-            // Use the unused retrieval helper function
-            test_location_retrieval(&app, &location_id).await;
-
-            // Use the unused update helper function
-            let _update_success = test_location_update(&app, &location_id).await;
-
-            // Use the unused deletion helper function
-            let _delete_success = test_location_deletion(&app, &location_id).await;
-
-            println!("✅ Complete location lifecycle test passed using helper functions");
-        }
-        Err(error) => {
-            println!("📋 Location lifecycle test failed: {error}");
-            // Test still passes - documents that the API may not be fully implemented
-        }
-    }
-}
-
-#[tokio::test]
-async fn test_multiple_location_operations() {
-    let app = setup_test_app().await;
-
-    // Create a test project first
-    let project_id = create_test_project(&app).await;
-    let project_id_str = project_id.to_string();
-
-    // Create multiple locations using the helper function
-    let mut location_ids = Vec::new();
-
-    for i in 1..=3 {
-        match create_test_location(&app, &project_id_str).await {
-            Ok((location_id, body)) => {
-                println!("✅ Location {i} created: {location_id}");
-                assert!(
-                    body["name"]
-                        .as_str()
-                        .unwrap()
-                        .contains("Test Location CRUD")
-                );
-                location_ids.push(location_id);
-            }
-            Err(error) => {
-                println!("📋 Location {i} creation failed: {error}");
-            }
-        }
-    }
-
-    // Test retrieval of all created locations
-    for (i, location_id) in location_ids.iter().enumerate() {
-        println!("Testing retrieval of location {}", i + 1);
-        test_location_retrieval(&app, location_id).await;
-    }
-
-    // Test updates on all locations
-    for (i, location_id) in location_ids.iter().enumerate() {
-        println!("Testing update of location {}", i + 1);
-        let _update_success = test_location_update(&app, location_id).await;
-    }
-
-    println!("✅ Multiple location operations test completed");
-}
-
-#[tokio::test]
-async fn test_location_error_handling() {
-    let app = setup_test_app().await;
-
-    // Test retrieval of non-existent location using helper
-    let fake_location_id = uuid::Uuid::new_v4().to_string();
-    println!("Testing retrieval of non-existent location: {fake_location_id}");
-
-    // This should not panic but handle the error gracefully
-    test_location_retrieval(&app, &fake_location_id).await;
-
-    // Test update of non-existent location using helper
-    println!("Testing update of non-existent location: {fake_location_id}");
-    let _update_success = test_location_update(&app, &fake_location_id).await;
-
-    // Test deletion of non-existent location using helper
-    println!("Testing deletion of non-existent location: {fake_location_id}");
-    let _delete_success = test_location_deletion(&app, &fake_location_id).await;
-
-    println!("✅ Location error handling test completed");
-}
-
-#[tokio::test]
-async fn test_location_helper_functions_consistency() {
-    let app = setup_test_app().await;
-
-    // Create a test project first
-    let project_id = create_test_project(&app).await;
-    let project_id_str = project_id.to_string();
-
-    // Create a location using the helper
-    match create_test_location(&app, &project_id_str).await {
-        Ok((location_id, create_body)) => {
-            println!("✅ Location created for consistency test: {location_id}");
-
-            // Verify the created location has all expected fields
-            assert!(create_body["id"].is_string());
-            assert!(create_body["name"].is_string());
-            assert!(create_body["comment"].is_string());
-            assert!(create_body["created_at"].is_string());
-            assert!(create_body["last_updated"].is_string());
-
-            // Test that retrieval helper works with the created location
-            test_location_retrieval(&app, &location_id).await;
-
-            // Verify the location can be updated and retrieved again
-            let _update_success = test_location_update(&app, &location_id).await;
-            test_location_retrieval(&app, &location_id).await;
-
-            println!("✅ Helper functions consistency test passed");
-        }
-        Err(error) => {
-            println!("📋 Consistency test skipped due to creation failure: {error}");
-        }
-    }
-}
