+++ src/routes/experiments/mod.rs
@@ -1,5 +1,6 @@
 pub mod excel_upload;
 pub mod models;
+pub mod models_old;
 pub mod phase_transitions;
 pub mod services;
 pub mod temperatures;
