+++ src/routes/experiments/models.rs
@@ -1,5 +1,3 @@
-//! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.8
-
 use sea_orm::entity::prelude::*;
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
@@ -25,49 +23,49 @@ pub struct Model {
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
 pub enum Relation {
-    #[sea_orm(has_many = "super::regions::Entity")]
+    #[sea_orm(has_many = "crate::routes::trays::regions::models::Entity")]
     Regions,
-    #[sea_orm(has_many = "super::s3_assets::Entity")]
+    #[sea_orm(has_many = "crate::routes::assets::models::Entity")]
     S3Assets,
-    #[sea_orm(has_many = "super::temperature_readings::Entity")]
+    #[sea_orm(has_many = "crate::routes::experiments::temperatures::models::Entity")]
     TemperatureReadings,
     #[sea_orm(
-        belongs_to = "super::tray_configurations::Entity",
+        belongs_to = "crate::routes::trays::configurations::models::Entity",
         from = "Column::TrayConfigurationId",
-        to = "super::tray_configurations::Column::Id",
+        to = "crate::routes::trays::configurations::models::Column::Id",
         on_update = "NoAction",
         on_delete = "NoAction"
     )]
     TrayConfigurations,
-    #[sea_orm(has_many = "super::well_phase_transitions::Entity")]
+    #[sea_orm(has_many = "crate::routes::experiments::phase_transitions::models::Entity")]
     WellPhaseTransitions,
 }
 
-impl Related<super::regions::Entity> for Entity {
+impl Related<crate::routes::trays::regions::models::Entity> for Entity {
     fn to() -> RelationDef {
         Relation::Regions.def()
     }
 }
 
-impl Related<super::s3_assets::Entity> for Entity {
+impl Related<crate::routes::assets::models::Entity> for Entity {
     fn to() -> RelationDef {
         Relation::S3Assets.def()
     }
 }
 
-impl Related<super::temperature_readings::Entity> for Entity {
+impl Related<crate::routes::experiments::temperatures::models::Entity> for Entity {
     fn to() -> RelationDef {
         Relation::TemperatureReadings.def()
     }
 }
 
-impl Related<super::tray_configurations::Entity> for Entity {
+impl Related<crate::routes::trays::configurations::models::Entity> for Entity {
     fn to() -> RelationDef {
         Relation::TrayConfigurations.def()
     }
 }
 
-impl Related<super::well_phase_transitions::Entity> for Entity {
+impl Related<crate::routes::experiments::phase_transitions::models::Entity> for Entity {
     fn to() -> RelationDef {
         Relation::WellPhaseTransitions.def()
     }
