+++ src/routes/samples/models.rs
@@ -1,8 +1,25 @@
-//! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.8
-
-use super::sea_orm_active_enums::SampleType;
 use sea_orm::entity::prelude::*;
+use serde::{Deserialize, Serialize};
+use utoipa::ToSchema;
 
+#[derive(
+    Debug, Clone, PartialEq, Eq, EnumIter, DeriveActiveEnum, ToSchema, Serialize, Deserialize,
+)]
+#[sea_orm(rs_type = "String", db_type = "Enum", enum_name = "sample_type")]
+pub enum SampleType {
+    #[sea_orm(string_value = "bulk")]
+    #[serde(rename = "bulk")]
+    Bulk,
+    #[sea_orm(string_value = "filter")]
+    #[serde(rename = "filter")]
+    Filter,
+    #[sea_orm(string_value = "procedural_blank")]
+    #[serde(rename = "procedural_blank")]
+    ProceduralBlank,
+    #[sea_orm(string_value = "pure_water")]
+    #[serde(rename = "pure_water")]
+    PureWater,
+}
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
 #[sea_orm(table_name = "samples")]
 pub struct Model {
@@ -42,24 +59,24 @@ pub struct Model {
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
 pub enum Relation {
     #[sea_orm(
-        belongs_to = "super::locations::models::Entity",
+        belongs_to = "crate::routes::locations::models::Entity",
         from = "Column::LocationId",
-        to = "super::locations::models::Column::Id",
+        to = "crate::routes::locations::models::Column::Id",
         on_update = "NoAction",
         on_delete = "NoAction"
     )]
     Locations,
-    #[sea_orm(has_many = "super::treatments::models::Entity")]
+    #[sea_orm(has_many = "crate::routes::treatments::models::Entity")]
     Treatments,
 }
 
-impl Related<super::locations::models::Entity> for Entity {
+impl Related<crate::routes::locations::models::Entity> for Entity {
     fn to() -> RelationDef {
         Relation::Locations.def()
     }
 }
 
-impl Related<super::treatments::models::Entity> for Entity {
+impl Related<crate::routes::treatments::models::Entity> for Entity {
     fn to() -> RelationDef {
         Relation::Treatments.def()
     }
