+++ src/routes/trays/wells/models.rs
@@ -17,24 +17,24 @@ pub struct Model {
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
 pub enum Relation {
     #[sea_orm(
-        belongs_to = "super::trays::Entity",
+        belongs_to = "crate::routes::trays::models::Entity",
         from = "Column::TrayId",
-        to = "super::trays::Column::Id",
+        to = "crate::routes::trays::models::Column::Id",
         on_update = "NoAction",
         on_delete = "NoAction"
     )]
     Trays,
-    #[sea_orm(has_many = "super::well_phase_transitions::Entity")]
+    #[sea_orm(has_many = "crate::routes::experiments::phase_transitions::models::Entity")]
     WellPhaseTransitions,
 }
 
-impl Related<super::trays::Entity> for Entity {
+impl Related<crate::routes::trays::models::Entity> for Entity {
     fn to() -> RelationDef {
         Relation::Trays.def()
     }
 }
 
-impl Related<super::well_phase_transitions::Entity> for Entity {
+impl Related<crate::routes::experiments::phase_transitions::models::Entity> for Entity {
     fn to() -> RelationDef {
         Relation::WellPhaseTransitions.def()
     }
