+++ src/routes/treatments/models.rs
@@ -1,7 +1,6 @@
-//! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.8
-
-use super::sea_orm_active_enums::TreatmentName;
 use sea_orm::entity::prelude::*;
+use serde::{Deserialize, Serialize};
+use utoipa::ToSchema;
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
 #[sea_orm(table_name = "treatments")]
@@ -20,28 +19,44 @@ pub struct Model {
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
 pub enum Relation {
-    #[sea_orm(has_many = "super::regions::Entity")]
+    #[sea_orm(has_many = "crate::routes::trays::regions::models::Entity")]
     Regions,
     #[sea_orm(
-        belongs_to = "super::samples::Entity",
+        belongs_to = "crate::routes::samples::models::Entity",
         from = "Column::SampleId",
-        to = "super::samples::Column::Id",
+        to = "crate::routes::samples::models::Column::Id",
         on_update = "NoAction",
         on_delete = "NoAction"
     )]
     Samples,
 }
 
-impl Related<super::regions::Entity> for Entity {
+impl Related<crate::routes::trays::regions::models::Entity> for Entity {
     fn to() -> RelationDef {
         Relation::Regions.def()
     }
 }
 
-impl Related<super::samples::Entity> for Entity {
+impl Related<crate::routes::samples::models::Entity> for Entity {
     fn to() -> RelationDef {
         Relation::Samples.def()
     }
 }
 
 impl ActiveModelBehavior for ActiveModel {}
+
+#[derive(
+    Debug, Clone, PartialEq, Eq, EnumIter, DeriveActiveEnum, ToSchema, Serialize, Deserialize,
+)]
+#[sea_orm(rs_type = "String", db_type = "Enum", enum_name = "treatment_name")]
+pub enum TreatmentName {
+    #[sea_orm(string_value = "none")]
+    #[serde(rename = "none")]
+    None,
+    #[sea_orm(string_value = "heat")]
+    #[serde(rename = "heat")]
+    Heat,
+    #[sea_orm(string_value = "h2o2")]
+    #[serde(rename = "h2o2")]
+    H2o2,
+}
